refactor: 🛠️ 移除日历和会议页面，更新消息页面为即将上线状态，替换图标和导航项
