+++ src/app/dashboard/user/docs/page.tsx
@@ -1,0 +1,5 @@
+import ComingSoonPage from '@/components/working';
+
+export default function SharedDocumentsPage() {
+  return <ComingSoonPage />;
+}
