+++ src/app/dashboard/workflow/page.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import ComingSoonPage from '@/components/working';
 
-const Messages = () => {
+const Workflow = () => {
   return <ComingSoonPage />;
 };
 
-export default Messages;
+export default Workflow;
